mcp-client/pyproject.toml
[project]
name = "mcp-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "mcp>=1.22.0",
    "notebook>=7.5.0",
    "openai>=2.8.1",
    "python-dotenv>=1.2.1",
]
